feat(operation): decimal irrigation duration (Double) + newest-first schedule list
- durationSeconds: Integer -> Double across the entity, schedule service, and
  scheduler so the duration setting accepts decimals (e.g. 0.5, 2.5 min) and the
  partial water amount on stop is computed from precise elapsed seconds (no
  integer truncation). The edge command still rounds to whole seconds for sleep.
- ManualIrrigationTab: duration input allows decimals (step 0.5); the schedule
  list now sorts by scheduledTime descending so newly created schedules show on top.

+++ cassavaBE/src/main/java/com/example/demo/entity/MongoEntity/IrrigationSchedule.java
@@ -24,7 +24,7 @@ public class IrrigationSchedule {
     private String fieldId;
     private Integer valveId;
     private Date scheduledTime;
-    private Integer durationSeconds;
+    private Double durationSeconds;
     private Double amount;
     private String userName;
     private Status status;
@@ -50,8 +50,8 @@ public class IrrigationSchedule {
     public Date getScheduledTime() { return scheduledTime; }
     public void setScheduledTime(Date scheduledTime) { this.scheduledTime = scheduledTime; }
 
-    public Integer getDurationSeconds() { return durationSeconds; }
-    public void setDurationSeconds(Integer durationSeconds) { this.durationSeconds = durationSeconds; }
+    public Double getDurationSeconds() { return durationSeconds; }
+    public void setDurationSeconds(Double durationSeconds) { this.durationSeconds = durationSeconds; }
 
     public Double getAmount() { return amount; }
     public void setAmount(Double amount) { this.amount = amount; }
